[BEAM-1218] Move GCP specific IO into separate module
