+++ sdks/python/apache_beam/io/google_cloud_platform/datastore/v1/datastoreio_test.py
@@ -23,12 +23,12 @@ from google.protobuf import timestamp_pb2
 from googledatastore import helper as datastore_helper
 from mock import MagicMock, call, patch
 
-from apache_beam.io.datastore.v1 import fake_datastore
-from apache_beam.io.datastore.v1 import helper
-from apache_beam.io.datastore.v1 import query_splitter
-from apache_beam.io.datastore.v1.datastoreio import _Mutate
-from apache_beam.io.datastore.v1.datastoreio import ReadFromDatastore
-from apache_beam.io.datastore.v1.datastoreio import WriteToDatastore
+from apache_beam.io.google_cloud_platform.datastore.v1 import fake_datastore
+from apache_beam.io.google_cloud_platform.datastore.v1 import helper
+from apache_beam.io.google_cloud_platform.datastore.v1 import query_splitter
+from apache_beam.io.google_cloud_platform.datastore.v1.datastoreio import _Mutate
+from apache_beam.io.google_cloud_platform.datastore.v1.datastoreio import ReadFromDatastore
+from apache_beam.io.google_cloud_platform.datastore.v1.datastoreio import WriteToDatastore
 
 
 class DatastoreioTest(unittest.TestCase):
