+++ sdks/python/apache_beam/io/google_cloud_platform/datastore/v1/helper.py
