+++ sdks/python/apache_beam/io/google_cloud_platform/gcsio.py
